sample-app/src/api-gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
python-json-logger
